Experimental pseudo-chess support.
Support for boards smaller than 8x8, for chess where the king is not
special and can be captured like any other piece, and blockers/holes
in the board to which no piece can move, and through which no pieces
can attack.

Only works in Clojure so far, not in ClojureScript.